[ALPS03498200] DPE: Fix kernel build error
[Detail] Fix kernel build error of DPE

MTK-Commit-Id: 532f8bb296485fee86511199585629c6e4872dde

CR-Id: ALPS03498200
Feature: Build System

+++ drivers/misc/mediatek/cameraisp/src/mt6757/camera_dpe.c
@@ -1154,7 +1154,7 @@ static signed int ConfigDVEHW(struct DPE_DVEConfig *pDveConfig)
 {
 #ifdef DPE_USE_GCE
 		struct cmdqRecStruct *handle;
-		uint64_t engineFlag = (1L << CMDQ_ENG_DPE);
+		uint64_t engineFlag = (uint64_t)(1LL << CMDQ_ENG_DPE);
 #endif
 
 	if (DPE_DBG_DBGLOG == (DPE_DBG_DBGLOG & DPEInfo.DebugMask)) {
@@ -1560,7 +1560,7 @@ static signed int ConfigWMFEHW(struct DPE_WMFEConfig *pWmfeCfg)
 {
 #ifdef DPE_USE_GCE
 		struct cmdqRecStruct *handle;
-		uint64_t engineFlag = (1L << CMDQ_ENG_DPE);
+		uint64_t engineFlag = (uint64_t)(1LL << CMDQ_ENG_DPE);
 #endif
 	unsigned int i = 0;
 	unsigned int wmfe_mask_value = 0;
